+++ packages/saved-views-react/src/experimental.ts
@@ -5,6 +5,3 @@
 export { createSavedViewOptions, type CreateSavedViewOptionsParams } from "./SavedViewTile/SavedViewOptions.js";
 export { SavedViewsExpandableBlockWidget } from "./SavedViewsWidget/SavedViewsExpandableBlockWidget.js";
 export { SavedViewsFolderWidget } from "./SavedViewsWidget/SavedViewsFolderWidget.js";
-export { applyExtensionsToViewport } from "./translation/SavedViewTranslation.js";
-export { createViewState } from "./createViewState.js";
-export { ModelCategoryOverrideProvider } from "./ModelCategoryOverrideProvider.js";